Add result step
