+++ src/components.d.ts
@@ -13,8 +13,11 @@ declare module '@vue/runtime-core' {
     FindTheWayOut: typeof import('./components/Games/FindTheWayOut.vue')['default']
     Game: typeof import('./components/Games/GamesUI/Game.vue')['default']
     GamePresentation: typeof import('./components/Games/GamePresentation/GamePresentation.vue')['default']
+    GustaveBubble: typeof import('./components/Games/IAGame/GustaveBubble.vue')['default']
     Home: typeof import('./views/Home.vue')['default']
     IAGame: typeof import('./components/Games/IAGame/IAGame.vue')['default']
+    IAGameResultStep: typeof import('./components/Games/IAGame/IAGameResultStep.vue')['default']
+    IAGameTrainingStep: typeof import('./components/Games/IAGame/IAGameTrainingStep.vue')['default']
     IIcBaselineDraw: typeof import('~icons/ic/baseline-draw')['default']
     IIcBaselineNetworkCheck: typeof import('~icons/ic/baseline-network-check')['default']
     IIcBaselineSecurity: typeof import('~icons/ic/baseline-security')['default']
